+++ src/homework6/Domestic.java
@@ -4,7 +4,7 @@ public class Domestic extends Animal {
     private String name;
     private boolean isVaccinated;
 
-    public Domestic(int age, int weight, String color, String name, boolean isVaccinated){
+    public Domestic(int age, double weight, String color, String name, boolean isVaccinated){
         super(age,weight,color);
         this.name = name;
         this.isVaccinated = isVaccinated;
