+++ src/homework6/Wild.java
@@ -3,7 +3,7 @@ package homework6;
 public class Wild extends Animal {
     private boolean isPredator;
 
-    public Wild(int age, int weight, String color,boolean isPredator) {
+    public Wild(int age, double weight, String color,boolean isPredator) {
         super(age, weight, color);
         this.isPredator = isPredator;
     }
